Product anzahl float;Products add wert

+++ anic/üben/test/Product.java
@@ -2,5 +2,5 @@
 
 public class Product
 {
-	private int anzahl;
+	private float anzahl;
 }
